% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomicAssigment.R
\name{annotateContigs.taxonomy}
\alias{annotateContigs.taxonomy}
\title{Assigns}
\usage{
annotateContigs.taxonomy(blast2lca)
}
\arguments{
\item{blast2lca}{output file showing taxonomic assignment}
}
\description{
annotateContigs
}

