% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomicAssigment.R
\name{blast2lcaIn}
\alias{blast2lcaIn}
\title{blast2lcaIN
reads in and executes perl}
\usage{
blast2lcaIn(blast2lca, targetFile = "blast2lca.csv", keepFile = FALSE)
}
\arguments{
\item{blast2lca}{output file showing taxonomic assignment}

\item{targetFile}{path to file which will store the intermediate output from the perl executable}

\item{keepFile}{to keep file or not}
}
\description{
blast2lcaIN
reads in and executes perl
}

