% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intervals.R
\name{buildIntervals}
\alias{buildIntervals}
\title{Builds intervals based on a stringSet Object}
\usage{
buildIntervals(stringSet, minIntervalSize)
}
\arguments{
\item{stringSet}{string Object from Biostrings}

\item{minIntervalSize}{the minimal size of the contig}
}
\value{
A GRanges object from \code{GenomicRanges} showing intervals
}
\description{
\code{buildIntervals} returns GRanges object with the ranges of contigs spanning the GLOBAL MSA
}
\details{
Takes in a DNAStringSet object from \code{Biostrings} generated from reading in the fasta MSA of contigs generated from pAss pipeline
}

