% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countNXX.R
\name{countInside}
\alias{countInside}
\title{counts the rolling abundance scores}
\usage{
countInside(countDF, abundanceCol, perc = perc)
}
\arguments{
\item{countDF}{df from the countNXX function}

\item{koID}{the KO id}
}
\description{
\code{countInside}
}

