% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countNXX.R
\name{countNXX}
\alias{countNXX}
\title{Assigns the a NXX to KOs}
\usage{
countNXX(countDF, perc = seq(0.1, 1, 0.05), abundanceCol = "rpkm_cDNA")
}
\arguments{
\item{countDF}{input data.frame}

\item{perc}{the percentages given as a sequence of}

\item{abundanceCol}{abundance}
}
\value{
a data.frame obj containing the NXX values.
We define NXX as metric for describing the number of genes required
to satisfy X% total expression within the KO
}
\description{
\code{countNXX} takes the RPKM expression value and returns a NXX value describing the number
of contigs required to account for X% of expression eg. 50% of expression
}

