% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DBprep.R
\name{DBprep}
\alias{DBprep}
\title{DBprep prepares input for NEO4J database}
\usage{
DBprep(theDIR, allKODF)
}
\arguments{
\item{allKODF}{data.frameObj which holds all count data}

\item{dir}{the directory to store the files}
}
\description{
DBprep takes the output from the mapBlat pipeline , which maps reads to the max diversity region
and prepares it as node, rel directories
for input by the omics package,
}

