% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countNXX.R
\name{distill}
\alias{distill}
\title{Returns the contigs and their assoc cDNA rpkms required for the KO to achieve N percentage of 100% expression}
\usage{
distill(countDF, nxxDF, koID, nxx)
}
\arguments{
\item{countDF}{contig table}

\item{nxxDF}{output data.frame after merging}
}
\description{
Returns the contigs and their assoc cDNA rpkms required for the KO to achieve N percentage of 100% expression
}

