% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intervals.R
\name{extendRange}
\alias{extendRange}
\title{Extends MAX diversity region to capture mRNAs mapping to ends of reads when executing 100 percent identity read matches}
\usage{
extendRange(spanningGR, truncGR, insideGR, mRNAsize, msaLength)
}
\arguments{
\item{spanningGR}{the spanning Genomicranges obj}

\item{truncGR}{the truncated Genomicranges obj}

\item{insideGR}{the inside Genomicranges obj}

\item{mRNAsize}{the size of the mRNA}

\item{msaLength}{the MSA length}
}
\value{
list
}
\description{
\code{extendRange} to allow mapping using Blat. Ie. Full map to mRNAs which touch the ends of the contigs
}
\details{
Function builds the range
}

