% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intervals.R
\name{extractLoc}
\alias{extractLoc}
\title{extractLoc: extracts MSA information from the header of the FASTA files}
\usage{
extractLoc(fileName)
}
\arguments{
\item{fileName}{the name of the file to inspect}
}
\value{
A list with the start and end location
}
\description{
takes info in header of FASTA output from pAss pipeline and generates and location of the MAX Diversity region
}

