% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{make.map.table}
\alias{make.map.table}
\title{make.map.table takes the output from map function (mapping reads onto contigs) and}
\usage{
make.map.table(contigs, reads, datype = "cDNA")
}
\arguments{
\item{overlaps}{output from the findOverlaps function from IRanges}
}
\value{
a data.frame object with the names of contigs and the number of reads mapped to them
}
\description{
takes the table which generates the overlap output back to data.frame
}

