% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.R
\name{map}
\alias{map}
\title{Map: maps gDNA and mRNA to contigs}
\usage{
map(fq, fa, blatOutputFile, contigs)
}
\arguments{
\item{fq}{vector of FASTQ files; read1, read2 or single read}

\item{fa}{FASTA file to store}

\item{blatOutputFile}{the output file to store results of blat}

\item{fasta}{file containing contigs}
}
\value{
a data.frame containing the readID the match and the blat information
}
\description{
Convert converts FASTQ input to FASTA format using perl Bio::SeqIO.
Runs BLAT in fast mode and outputs the matches
}

