% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapper}
\alias{mapper}
\title{Map: maps gDNA and mRNA to contigs}
\usage{
mapper(blatOutputFile)
}
\arguments{
\item{blatOutputFile}{the output file to store results of blat}
}
\value{
a data.frame containing the readID the match and the blat information
}
\description{
Convert converts FASTQ input to FASTA format using perl Bio::SeqIO.
Runs \code{BLAT -fastMap} in fast mode and outputs the matches.
Matches are redundant ie. Reads are matched to more than one contig.
}

