% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intervals.R
\name{spanningOrNot}
\alias{spanningOrNot}
\title{Determines nature of contigs wrt to Max Diversity Region; is this spanning, non-spanning, entirely captured inside}
\usage{
spanningOrNot(gr, startLOC, endLOC)
}
\arguments{
\item{gr}{A GRanges Object}

\item{startLOC}{the start location for the region of interest}

\item{endLOC}{the end location for the region of interest}
}
\value{
Returns a list of GRanges Objects which are 1. spanning, 2. non-spanning or entirely inside
}
\description{
\code{spanningOrNot} Figures out the way the contigs map to the max diversity regions
}
\details{
This function is part of a series
}

