% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upMSA.R
\docType{data}
\name{upMSA}
\alias{upMSA}
\title{MSA of contigs against the MSA of reference sequences}
\format{A DNAStringSet instance of length 1600
\describe{
    \item{Dependencies}{Requires the Biostrings package from Bioconductor}
    \item{Names}{Names of the contigs which the sequences are aligned to}
}}
\usage{
upMSA
}
\description{
1600 contigs aligned against all references sequences in
ko:K00001, alcohol dehydrogenase
The best match chosen then reMSA according to MSA of
reference sequences.
}
\keyword{datasets}

