% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intervals.R
\name{extendRange}
\alias{extendRange}
\title{Extends MAX diversity region to capture mRNAs mapping to ends of reads when executing 100% percent identity read matches}
\usage{
extendRange(spanningGR, truncGR, insideGR, mRNAsize, msaLength)
}
\arguments{
\item{spanningGR}{}

\item{truncGR}{}

\item{insideGR}{}

\item{mRNAsize}{}

\item{msaLength}{}
}
\value{
list
}
\description{
extendRange to allow mapping using Blat. Ie. 100% map to mRNAs which touch the ends of the contigs
}

